update: 重构底层架构与多组支持，优化界面与性能
- **底层架构重写**：
  - FaceConfig 从单组 ID 升级为多组 ID 集合，一个面可同时属于多个物流组。
  - 反向链接索引重构为按需从 linkedNodes 真源全量重建，消除索引漂移风险。
  - 清理未使用的全局缓存、死方法和冗余配置项，统一 Key 编码至 LogisticsNode。
  - 线程安全加固：dirtyFaceKeys/dirtyContainerKeys 加锁保护，消除并发竞争。

- **多组管理**：
  - 移除模式必须选取组才能操作，按组精确移除不再误删其他组链接。
  - 重命名服务适配多组，使用 removeGroupId + addGroupId 替代破坏性 setGroupId。
  - 节点注册层与客户端数据适配多组查询。

- **界面与交互优化**：
  - 配置器类型选择从单选轮询改为多选 toggle（点击切换启用/禁用）。
  - 分发策略和新增的提取模式按钮支持左键下一个、右键上一个。
  - 优先级输入框移除 ±9999 限制，仅允许数字和负号。
  - 过滤器界面添加手持物品/流体的左键物品、右键流体快速标记提示。

- **性能与修复**：
  - 渲染性能优化：renderFlows 复用 activeNodes 缓存，消除重复查找。
  - 传输性能优化：selectTargets 复用 FaceConfig 查找结果。
  - 修复容器配置模式对多方块结构（Mekanism 阀门、Create 壳体等）的误判。
  - 修复配置热重载后已插入升级卡不生效的问题（通过 configGeneration 计数器）。
  - 最大传输量上限从 int 升级为 long，支持整合包极端数值。

- **代码清理与依赖更新**：
  - 删除未使用的 PneumaticCraft 气压与热量传输支持。
  - 更新 Mekanism 依赖版本，添加 Mekanism Extras 和 More Machines 运行时依赖。
  - 补充遗漏翻译键，删除未使用的翻译键，配置类 tooltip 全覆盖。

+++ src/main/java/com/coobird/staticlogistics/util/LogisticsCalculator.java
@@ -46,7 +46,7 @@ public final class LogisticsCalculator {
      */
     public static long getTransferLimit(ContainerConfig container, TransferType type) {
         if (container == null) {
-            return Math.min(type.getBaseStackSize(), SLConfig.getMaxTransferLimit());
+            return Math.min((long) type.getBaseStackSize(), SLConfig.getMaxTransferLimit());
         }
         long stackMult = getStackMultiplier(container);
         long limit = (long) type.getBaseStackSize() * stackMult;
